+++ packages/react/src/index.ts
@@ -10,6 +10,7 @@ export * from './components/animate-grow/index.js';
 export * from './components/animate-headline/index.js';
 export * from './components/animate-lighting/index.js';
 export * from './components/animate-marquee/index.js';
+export * from './components/animate-reveal/index.js';
 export * from './components/animate-rotate/index.js';
 export * from './components/animate-slide/index.js';
 export * from './components/animate-typing/index.js';